Added Jupyter notebook over scrap project
In the future I will be moving my work in jupyter notebook over to a script for upload into Git Hub. Considering this is my first upload in some time, I will stick to my intentions, which is to work on a data project and upload it to GitHub. This goes against my perfectionist tendencies which have held me back from learning. If you are an employer I would not recommend looking at this scrap project as an indication of what I am capable as instead look at it as a scrap piece of paper you have found on my desk .. enjoy :)